requirements.txt
jsonpickle==1.4.1
json5==0.8.4
Flask==1.1.1
Flask-Classful==0.14.2
p2pnetwork==0.0.3
requests==2.22.0
# pycrypto==2.6.1
pycryptodome
uuid